Structurer: Move final DREAM cleanup into DreamStructurer
